OS X: compile util
This is a rollup of commits authored by Adar, JD, and me. These changes allow
the util module to be compiled on OS X.

+++ src/kudu/util/env.h
@@ -309,7 +309,7 @@ class RandomAccessFile {
                       uint8_t *scratch) const = 0;
 
   // Returns the size of the file
-  virtual Status Size(uint64_t *size) const = 0;
+  virtual Status Size(size_t *size) const = 0;
 
   // Returns the filename provided when the RandomAccessFile was constructed.
   virtual const std::string& filename() const = 0;
